+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Enhancement:** Improved `Views` report under `Visitor Insight` page.
 - **Enhancement:** Unset `Consent Integration` option upon integration deactivation.
 - **New:** Add Country Name to Visitor Info Flag Tooltip.
+- **New:** Added unread notification badge to Help Center menu.
 
 = 14.14 - 2025-06-01 =
 - **New:** Added a Help page inside the plugin to guide users.